Added LeftSection buttons
