+++ src/pages/admin/index.jsx
@@ -1,24 +1,10 @@
-import styles from "@/styles/Admin.module.css";
 import AuthCheck from "@/components/AuthCheck";
 import PostFeed from "@/components/PostFeed";
-import { UserContext } from "@/lib/context";
-import { firestore, auth } from "@/lib/firebase";
-import {
-  serverTimestamp,
-  query,
-  collection,
-  orderBy,
-  getFirestore,
-  setDoc,
-  doc,
-} from "firebase/firestore";
-
-import { useContext, useState } from "react";
-import { useRouter } from "next/router";
+import CreateNewPost from "@/components/CreateNewPost";
+import { auth } from "@/lib/firebase";
+import { query, collection, orderBy, getFirestore } from "firebase/firestore";
 
 import { useCollection } from "react-firebase-hooks/firestore";
-import kebabCase from "lodash.kebabcase";
-import toast from "react-hot-toast";
 
 export default function AdminPostsPage(props) {
   return (
@@ -33,9 +19,6 @@ export default function AdminPostsPage(props) {
 }
 
 function PostList() {
-  // const ref = firestore.collection('users').doc(auth.currentUser.uid).collection('posts');
-  // const query = ref.orderBy('createdAt');
-
   const ref = collection(
     getFirestore(),
     "users",
@@ -52,63 +35,5 @@ function PostList() {
     <>
       <PostFeed posts={posts} admin />
     </>
-  );
-}
-
-function CreateNewPost() {
-  const router = useRouter();
-  const { username } = useContext(UserContext);
-  const [title, setTitle] = useState("");
-
-  // Ensure slug is URL safe
-  const slug = encodeURI(kebabCase(title));
-
-  // Validate length
-  const isValid = title.length > 3 && title.length < 100;
-
-  // Create a new post in firestore
-  const createPost = async (e) => {
-    e.preventDefault();
-    const uid = auth.currentUser.uid;
-    const ref = doc(getFirestore(), "users", uid, "posts", slug);
-
-    // Tip: give all fields a default value here
-    const data = {
-      title,
-      slug,
-      uid,
-      username,
-      rating: 0,
-      published: false,
-      dish: "",
-      content: "",
-      createdAt: serverTimestamp(),
-      updatedAt: serverTimestamp(),
-      heartCount: 0,
-    };
-
-    await setDoc(ref, data);
-
-    toast.success("Post created! 🔥");
-
-    // Imperative navigation after doc is set
-    router.push(`/admin/${slug}`);
-  };
-
-  return (
-    <form onSubmit={createPost}>
-      <input
-        value={title}
-        onChange={(e) => setTitle(e.target.value)}
-        placeholder="Restaurant Name"
-        className={styles.input}
-      />
-      <p>
-        <strong></strong> {slug}
-      </p>
-      <button type="submit" disabled={!isValid} className="btn-green">
-        Create New Review
-      </button>
-    </form>
   );
 }
