+++ src/styles/LeftSection.module.css
@@ -5,6 +5,39 @@
   display: none;
 }
 
+.list {
+  margin-top: 100px;
+  display: flex;
+  flex-direction: column;
+  width: 100%;
+  align-items: center;
+  justify-content: center;
+}
+
+.item {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  justify-content: flex-start;
+  padding-left: 20px;
+  padding-top: 20px;
+  padding-bottom: 20px;
+  width: 100%;
+  height: 55px;
+  gap: 0.75rem;
+  font-size: 15px;
+  width: 75%;
+  border-radius: 10px;
+}
+
+.item:hover {
+  background-color: var(--color-bg);
+}
+
+.icon {
+  font-size: 30px;
+}
+
 /* 
     DESKTOP SCREEN
 */
@@ -12,8 +45,7 @@
   .left {
     display: flex;
     flex-direction: column;
-    align-items: center;
-    justify-content: center;
+    justify-content: flex-start;
     height: 100%;
     width: 20%;
     position: fixed;
